+++ middleware/contextualizers/enterprise-release-notes.js
@@ -1,6 +1,15 @@
+const semver = require('semver')
 const renderContent = require('../../lib/render-content')
 const patterns = require('../../lib/patterns')
-const enterpriseReleases = require('../../lib/enterprise-server-releases').supported
+const { all, firstReleaseNote, latest } = require('../../lib/enterprise-server-releases')
+
+// Display all GHES release notes, regardless of deprecation status,
+// starting with the first release notes in 2.20
+const supported = all.filter(release => {
+  return semver.gte(
+    semver.coerce(release), semver.coerce(firstReleaseNote)
+  ) && release !== '11.10.340'
+})
 
 /**
  * Turn { [key]: { notes, intro, date } }
@@ -17,10 +26,24 @@ function sortPatchKeys (release, version) {
         ...release[key]
       }
     })
+    // Filter out any deprecated patches
+    .filter(key => !key.deprecated)
   return keys
     .sort((a, b) => {
-      if (a.version > b.version) return -1
-      if (a.version < b.version) return 1
+      let aTemp = a.version
+      let bTemp = b.version
+
+      // There's an RC version here, so doing regular semver
+      // comparisons won't work. So, we'll convert the incompatible version
+      // strings to real semver strings, then compare.
+      const [aBase, aRc] = a.version.split('.rc')
+      if (aRc) aTemp = `${aBase}-rc.${aRc}`
+
+      const [bBase, bRc] = b.version.split('.rc')
+      if (bRc) bTemp = `${bBase}-rc.${bRc}`
+
+      if (semver.gt(aTemp, bTemp)) return -1
+      if (semver.lt(aTemp, bTemp)) return 1
       return 0
     })
 }
@@ -51,9 +74,9 @@ async function renderPatchNotes (patch, ctx) {
   return patch
 }
 
-module.exports = async (req, res, next) => {
+module.exports = async function enterpriseReleaseNotesContext (req, res, next) {
   // The `/release-notes` sub-path
-  if (!req.path.endsWith('/release-notes')) return next()
+  if (!(req.path.endsWith('/release-notes') || req.path.endsWith('/admin'))) return next()
 
   // ignore paths that don't have an enterprise version number
   if (!patterns.getEnterpriseServerNumber.test(req.path)) return next()
@@ -77,18 +100,25 @@ module.exports = async (req, res, next) => {
   req.context.releaseNotes = await Promise.all(patches.map(async patch => renderPatchNotes(patch, req.context)))
 
   // Put together information about other releases
-  req.context.releases = enterpriseReleases.map(version => {
+  req.context.releases = supported.map(version => {
     const ret = { version }
     if (!req.context.site.data['release-notes']) return ret
     const release = req.context.site.data['release-notes'][version.replace(/\./g, '-')]
     if (!release) return ret
     const patches = sortPatchKeys(release, version)
-    return { ...ret, patches }
+
+    const firstPreviousRelease = all[all.findIndex(v => v === version) + 1]
+    const secondPreviousRelease = all[all.findIndex(v => v === firstPreviousRelease) + 1]
+
+    return { ...ret, patches, firstPreviousRelease, secondPreviousRelease }
   })
 
-  const releaseIndex = enterpriseReleases.findIndex(release => release === requestedVersion)
-  req.context.nextRelease = enterpriseReleases[releaseIndex - 1]
-  req.context.prevRelease = enterpriseReleases[releaseIndex + 1]
+  const releaseIndex = supported.findIndex(release => release === requestedVersion)
+  req.context.nextRelease = supported[releaseIndex - 1]
+  req.context.prevRelease = supported[releaseIndex + 1]
+
+  req.context.latestPatch = patches[0].version
+  req.context.latestRelease = latest
 
   return next()
 }