+++ script/enterprise-server-releases/create-graphql-files.js
@@ -55,10 +55,15 @@ const objectsFile = path.join(graphqlStaticDir, 'prerendered-objects.json')
 const previews = require(previewsFile)
 const changes = require(changesFile)
 const objects = require(objectsFile)
+// The prerendered objects file for the "old version" contains hardcoded links with the old version number.
+// We need to update those links to include the new version to prevent a test from failing.
+const regexOldVersion = new RegExp(oldVersion, 'gi')
+const stringifiedObject = JSON.stringify(objects[oldVersionId])
+  .replace(regexOldVersion, newVersion)
 
 previews[newVersionId] = previews[oldVersionId]
 changes[newVersionId] = changes[oldVersionId]
-objects[newVersionId] = objects[oldVersionId]
+objects[newVersionId] = JSON.parse(stringifiedObject)
 
 // check that it worked
 if (!Object.keys(previews).includes(newVersionId)) {